Changed results response so that nextUri and afterId are always included
- Changed results response so that the `nextUri` and `afterId` fields
are always included in the response, even if there are no filtered
results, so that the client can request the next page of results
- Changed Results service methods return types to
`ResultsResponseModel`
- Removed ResultsEntity and ResultsDTO since they are not needed
- No longer need to store the `nextUri` and `afterId` into the
database, since they are directly returned with the request response
- Added `.isEmpty()` conditional checks for lists
