+++ src/main/java/ca/tunestumbler/api/io/entity/ResultsEntity.java
@@ -65,12 +65,6 @@ public class ResultsEntity implements Serializable {
 	@Column(nullable = false)
 	private String url;
 
-	@Column
-	private String nextUri;
-
-	@Column
-	private String afterId;
-
 	@Column()
 	private Long startId;
 
@@ -203,22 +197,6 @@ public class ResultsEntity implements Serializable {
 
 	public void setUrl(String url) {
 		this.url = url;
-	}
-
-	public String getNextUri() {
-		return nextUri;
-	}
-
-	public void setNextUri(String nextUri) {
-		this.nextUri = nextUri;
-	}
-
-	public String getAfterId() {
-		return afterId;
-	}
-
-	public void setAfterId(String afterId) {
-		this.afterId = afterId;
 	}
 
 	public Long getStartId() {
