+++ src/main/java/ca/tunestumbler/api/service/ResultsService.java
@@ -1,12 +1,10 @@
 package ca.tunestumbler.api.service;
 
-import java.util.List;
-
-import ca.tunestumbler.api.shared.dto.ResultsDTO;
 import ca.tunestumbler.api.shared.dto.UserDTO;
+import ca.tunestumbler.api.ui.model.response.ResultsResponseModel;
 
 public interface ResultsService {
-	List<ResultsDTO> fetchResults(UserDTO user, String orderBy);
+	ResultsResponseModel fetchResults(UserDTO user, String orderBy);
 
-	List<ResultsDTO> fetchNextResults(UserDTO user, String nextUri, String nextAfterId);
+	ResultsResponseModel fetchNextResults(UserDTO user, String nextUri, String nextAfterId);
 }
