+++ src/main/java/ca/tunestumbler/api/ui/model/response/ResultsResponseModel.java
@@ -6,6 +6,8 @@ import ca.tunestumbler.api.ui.model.response.results.ResultsObjectResponseModel;
 
 public class ResultsResponseModel {
 	private List<ResultsObjectResponseModel> results;
+	private String nextUri;
+	private String afterId;
 
 	public List<ResultsObjectResponseModel> getResults() {
 		return results;
@@ -13,6 +15,22 @@ public class ResultsResponseModel {
 
 	public void setResults(List<ResultsObjectResponseModel> results) {
 		this.results = results;
+	}
+
+	public String getNextUri() {
+		return nextUri;
+	}
+
+	public void setNextUri(String nextUri) {
+		this.nextUri = nextUri;
+	}
+
+	public String getAfterId() {
+		return afterId;
+	}
+
+	public void setAfterId(String afterId) {
+		this.afterId = afterId;
 	}
 
 }
